Move HealthCheckSettings class to a separate file
