+++ XenServerHealthCheck/RequestUploadTask.cs
@@ -32,6 +32,7 @@ using System;
 using System.Collections.Generic;
 using XenAdmin.Network;
 using XenAPI;
+using XenAdmin.Model;
 
 namespace XenServerHealthCheck
 {
